Refactored, Maven, comments, how tos
