+++ HOW-TO-ADD-RDMS.md
@@ -1,0 +1,12 @@
+These are the instructions to add a new RDBMS.
+
+# Java
+
+There is a FIXME in each part of the Javadocs where modifications should be
+done for a new RDBMS.
+
+# List of changes
+
+ * jTPCC.java
+ * jTPCCConfig.java
+ * jTCPPConnection.java
