+++ HOW-TO-RUN-firebird.md
@@ -1,0 +1,32 @@
+
+Instructions for running BenchmarkSQL on Firebird
+-------------------------------------------------
+
+# Requirements.
+
+ToDo
+
+## Compile the BenchmarkSQL source code.
+
+ToDo
+
+# Create a database and the credentials to access.
+
+ToDo
+
+# Create the BenchmarkSQL configuration file.
+
+ToDo
+
+# Build the schema and initial database load.
+
+ToDo
+
+# Run the configured BenchmarkSQL
+
+ToDo
+
+# Scale the benchmark configuration
+
+ToDo
+
