+++ build.xml
@@ -28,7 +28,7 @@
 
   <target name="dist" depends="compile" >
     <mkdir dir="${dist}"/>
-    <jar jarfile="${dist}/BenchmarkSQL-6.devel.jar" basedir="${build}"/>
+    <jar jarfile="${dist}/BenchmarkSQL.jar" basedir="${build}"/>
   </target>
 
   <target name="clean" description="clean up" >
